Update of Weighting Efficiency calc
Update of the WE calc to include the user's production boosts in order to get the base production values.
For example, previously a player entering that he had a 40'000 FP production would get a value for 1% FP Boost of 400 FPs flat. However, 1% FP boost did not necessarily meant an increase of 400 FPs, as the user probably had FP% active in their city. Now, if a user has a 40'000 FP production and 183% FP boost, it will remove the increase of 183% before calculating the flat value of 1% FP.

Therefore we have more realistic values for efficiency.

The update also includes the addition of credits and acknowledgements at the bottom of the webpage.