Added icons
Added icons for Ally Room, Limited, Nbr of squares (Avg) and size.
Updated data_loader.py to prepare it to retrieve diamonds and blueprints, as well as for showing % of chance to get some rewards in other productions. Updated it to retrieve asset_id for further updates on building image assets.

Updated gitignore with some MD files to ignore.